Calc'ing lambert on the fly instead of using tex
Admittedly, I don't know whether this is actually faster. But I think it is. The extra dot product call in the merge pass is presumably a lot faster than a writing into *and* doing a texture lookup for each screen pixel.

Also, this reduces duplication as we no longer have to duplicate the lambertian code in both Static and Bone geometry vert shader.

Also, this welcomes my upcoming feature of making a "universal lambertian" in addition to the primary light source. Without this change, we'd have to make yet another texture write/lookup for each pixel.

Overall, this "inline calc" approach seems a lot simpler - and probably faster too.