Introduced better lighting with pri/sec lights
Before this, we only had one and only light, the shadow casting light. This is now called the primary light, and a new one was added: secondary light.

The secondary light is weak and doesn't produce shadow. Its purpose is only to give a hint of shape (via lambertian lighting) in the rare occations when the primary light is not visible (for instance when the sun has just set).