First stab at low res SSAO
Unfortunately, this does not work at the moment. The reason it shows up "relatively" fine when running, is that we're still running with full resolution blurring in merge shader.

The reason it doesn't work is that we're using a full resolution depth texture in the half resolution blur shader. It doesn't work very well.

We must somehow downsample the input textures for the SSAO passes. I'll find a solution :-)