Slight improvement Orwell API
This frees up the use of the word/concept "suite". It is better used as a collection of test for a certain group of logic/code. For instance, with this change I can in the future make a test suite with test groups with unit tests for a certain library, i.e. Quanta Test Suite.