Reverted Screen/ScreenSync solution
This solution turned out to be not as great as I initially thought. Several downsides:

1) Apple generally do not recommend using same context from different threads
2) Could not easily solve the problem that it would consume all CPU when OpenGL was hidden (because flushBuffer returns immediately when hidden)
3) I've come up with a better design that I'll implement eventually: All input and rendering on main thread, move out network.

So all in all it's just not worth keeping the solution as it caused problems *and* would be replaced eventually anyway.

The downside of replacing this right now, is that client networking will only be updated 60 times a second which isn't perfect - but good enough. This problem will be solved when implementation design mentioned in 3).