Extracted DB logic to own module
The idea behind this is we can use this from Client module. Without extracting like this, Client had to maintain its own DB system written from scratch.

The DB logic gets a wee bit more complex because of this generalization, but I think it's worth. At least it's worth a try. Let's see where it goes.