Actions: NodeCall
This fixes a stupid design choice. Before this change, NodeFlow properties required by action nodes were overwritten when a node called another node. This would for example mean that a SequenceNode lost the ability to call flow->getConfig() it is called flow->start(x) first.

Node specific properties should be per call, not per flow. With the introduction of NodeCall order is restored.