Removed Fixie::Num% operator
It is confusing because it behaves like an int modulus even though Fixie::Num doesn't represent an int.

If you want to do int modulus you should cast Fixie::Num to an int and then do it.

If you want to do fmod() style modulus, you should Fixie::Util::fmod() (and you'd have to code it first because it doesn't exist yet).